+++ pandoc.cabal
@@ -618,7 +618,8 @@ test-suite test-pandoc
                   containers >= 0.4.2.1 && < 0.6,
                   executable-path >= 0.0 && < 0.1,
                   zip-archive >= 0.2.3.4 && < 0.4,
-                  xml >= 1.3.12 && < 1.4
+                  xml >= 1.3.12 && < 1.4,
+                  Glob >= 0.7 && < 0.10
   if flag(old-locale)
      build-depends: old-locale >= 1 && < 1.1,
                     time >= 1.2 && < 1.5
@@ -669,6 +670,7 @@ test-suite test-pandoc
                   Tests.Writers.Muse
                   Tests.Writers.FB2
                   Tests.Writers.Powerpoint
+                  Tests.Writers.OOXML
   ghc-options:  -rtsopts -Wall -fno-warn-unused-do-bind -threaded
   default-language: Haskell98
